requirements/base.txt
Django==5.0
asgiref==3.11.1
django-environ==0.13.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0